Move tab view model to the framework
